requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
aiofiles==24.1.0
python-dotenv==1.0.1
